requirements.txt
altgraph==0.17.2
APScheduler==3.6.3
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
docxcompose==1.3.3
docxtpl==0.12.0
filelock==3.0.12
flake8==3.9.2
future==0.18.2
idna==3.2
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mccabe==0.6.1
objectpath==0.6.1
pefile==2021.9.3
Pillow==8.3.2
platformdirs==2.3.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
python-docx==0.8.11
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
telegram==0.0.1
terminaltables==3.1.10
tornado==6.1
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
virtualenv==20.7.2
